requirements.txt
tflite-runtime
numpy>=1.23.5
Pillow>=9.5.0
python-multipart 